purescript: use node-readline instead of simple-repl
I thought maybe something about simple-repl might be causing the
test integration issues, but this impl has the same problem: it is
necessary to log the result twice as well as use the --no-pty on
the make command for tests to work. Maybe due to the async nature
of node readline.