purescript: progress on immutable reader
The mal spec describes a stateful OO reader, and this is an attempt
at an immutable reader that is still roughly in line with the spec.
The read* functions take a reader at a particular position, and return
a new reader at the post-read position, plus the value produced by
the read.